+++ deployment/powershell/AddImage.ps1
@@ -400,7 +400,7 @@ elseif ((!$skip2019Images) -and ($progressCheck -ne "Complete")) {
                     $date = Get-Date -Format FileDate
                     # Temporarily hard coding to newest known working version
                     #$vhdVersion = "16.04.$date"
-                    $vhdVersion = "16.04.20191114"
+                    $vhdVersion = "16.04.20200318"
                 }
                 else {
                     $vhdVersion = ""
@@ -497,7 +497,7 @@ elseif ((!$skip2019Images) -and ($progressCheck -ne "Complete")) {
                         $azpkg.vhdPath = $downloadDetails.properties.osDiskImage.sourceBlobSasUri
                         # Temporarily hard coding to newest known working Ubuntu image
                         #$azpkg.vhdVersion = $downloadDetails.properties.version
-                        $azpkg.vhdVersion = "16.04.20191114"
+                        $azpkg.vhdVersion = "16.04.20200318"
                     }
                 }
                 <#